requirements.txt
amqp==2.5.2
anyjson==0.3.3
atomicwrites==1.3.0
attrs==19.3.0
billiard==3.6.1.0
bootstrap4==0.1.0
cairocffi==1.0.2
CairoSVG==2.4.2
celery==4.3.0
cffi==1.13.2
colorama==0.4.1
cssselect2==0.2.2
defusedxml==0.6.0
Django==2.2.7
django-appconf==1.0.3
django-celery-beat==1.5.0
django-celery-email==2.0.2
django-celery-results==1.1.2
django-compat==1.0.15
django-crispy-forms==1.8.0
django-debug-toolbar==2.0
django-timezone-field==3.1
Faker==2.0.3
gunicorn==19.9.0
html5lib==1.0.1
importlib-metadata==0.23
kombu==4.6.3
mixer==6.1.3
more-itertools==7.2.0
mysqlclient==1.4.4
Pillow==6.2.1
pluggy==0.13.0
pycparser==2.19
Pyphen==0.9.5
python-crontab==2.4.0
python-dateutil==2.8.1
python-memcached==1.59
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.3
tinycss2==1.0.2
vine==1.3.0
WeasyPrint==50
webencodings==0.5.1
zipp==0.6.0
